move and remove
