+++ embc-app/ClientApp/src/app/app.module.ts
@@ -16,7 +16,7 @@ import { SharedModule } from './shared/shared.module';
 import { RegistrationSummaryComponent } from './registration-summary/registration-summary.component';
 import { TaskNumberListComponent } from './provincial-admin/components/task-number-list/task-number-list.component';
 import { VolunteerListComponent } from './local-authority/components/volunteer-list/volunteer-list.component';
-import { RegistrationListComponent } from './registration-list/registration-list.component';
+import { RegistrationListComponent } from './volunteer/components/registration-list/registration-list.component';
 import { OrganizationListComponent } from './provincial-admin/components/organization-list/organization-list.component';
 import { TaskNumberMakerComponent } from './provincial-admin/components/task-number-maker/task-number-maker.component';
 import { VolunteerMakerComponent } from './volunteer-maker/volunteer-maker.component';
